Forbid deletion of workout exercise if it is not empty

+++ src/Service/Workout/WorkoutExerciseService.php
@@ -36,6 +36,10 @@ final readonly class WorkoutExerciseService
 
     public function delete(WorkoutExercise $workoutExercise): void
     {
+        if (!$workoutExercise->sets->isEmpty()) {
+            throw new BadRequestHttpException('Workout exercise has sets');
+        }
+
         $this->repository->delete($workoutExercise);
     }
